+++ simtools.gaml.extensions.physics/src/javax/vecmath/Color3b.java
@@ -139,9 +139,9 @@ public class Color3b extends Tuple3b implements java.io.Serializable {
      * @since vecmath 1.2
      */
     public final Color get() {
-	int r = (int)x & 0xff;
-	int g = (int)y & 0xff;
-	int b = (int)z & 0xff;
+	int r = x & 0xff;
+	int g = y & 0xff;
+	int b = z & 0xff;
 
 	return new Color(r, g, b);
     }
